requirements.txt
textblob
spacy
google-auth
google-auth-oauthlib 
google-api-python-client 
python-dotenv
